Disable automatic haddock upload on stack
While cabal has a "--haddocks-for-hackage" flag, stack does not.
This is an open issue since 2015.

See: https://github.com/commercialhaskell/stack/issues/737

It would probably be easy to implement this in releaser, but that is out of
scope for this commit.

This commit adds also a very rudimentary detectBuildSystem utility. It
might become useful in general.